MA-419 Implement metrics feature with data model and UI components; update BodyScaleEntryEntity to use Double for metrics

+++ Android/app/src/main/java/com/greatergoods/meapp/data/storage/db/entity/entry/BodyScaleEntryEntity.kt
@@ -23,9 +23,9 @@ data class BodyScaleEntryEntity(
     @PrimaryKey
     val id: Long,
     val weight: Double,
-    val bodyFat: Int?,
-    val muscleMass: Int?,
-    val water: Int?,
-    val bmi: Int?,
+    val bodyFat: Double?,
+    val muscleMass: Double?,
+    val water: Double?,
+    val bmi: Double?,
     val source: String?,
 )
